Refactor a bit to clean up
