+++ code_data/test_minimized/astroid.brain.brain_numpy_ndarray.py
