+++ code_data/test_minimized/idlelib.idle_test.test_hyperparser.py
