+++ code_data/test_minimized/prompt_toolkit.styles.named_colors.py
