+++ code_data/test_minimized/setuptools.config.py
