+++ code_data/test_minimized/turtledemo.minimal_hanoi.py
